#620: Finish refactoring of app plugin
