+++ lib/lando.js
@@ -179,6 +179,17 @@ module.exports = _.memoize(function(config) {
     Promise: require('./promise'),
 
     /**
+     * The scan module.
+     *
+     * Contains a URL scanner
+     *
+     * @since 3.0.0
+     * @memberof lando
+     * @see {@link shell.md}
+     */
+    scanUrls: require('./scan')(log),
+
+    /**
      * The shell module.
      *
      * Contains helpful methods to parse and execute commands.
