+++ plugins/lando-app/index.js
@@ -191,6 +191,27 @@ module.exports = function(lando) {
 
     });
 
+    // Also gather info on a start
+    app.events.on('post-start', 1, function() {
+      return lando.app.info(app);
+    });
+
+    // Scan our URLs and add a list of them and their status to the app
+    app.events.on('post-start', 9, function() {
+
+      // Get app URLs
+      var urls = _.filter(_.flatMap(app.info, 'urls'), _.identity);
+
+      // Scan the urls
+      return lando.scanUrls(urls, {max: 17})
+
+      // Add our URLS to the app
+      .then(function(urls) {
+        app.urls = urls;
+      });
+
+    });
+
   });
 
 };
